Move DockerConfigPlugin to its own package
